% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a13.eq.ac.R
\name{a13.eq.ac}
\alias{a13.eq.ac}
\title{function used within the multiplier effort analysis}
\usage{
a13.eq.ac(compteur, ET_Main, biom.mf, fish.m.ac, TopD, FormD, range.TLpred)
}
\arguments{
\item{compteur}{counter}

\item{ET_Main}{output of the create.ETmain function}

\item{biom.mf}{Effort multiplier applied on biomass}

\item{fish.m.ac}{Effort multiplier applied on natural mortality}

\item{TopD}{Parameters of the formula}

\item{FormD}{Parameters of the formula}

\item{range.TLpred}{Range of predators trophic level}
}
\description{
function used within the multiplier effort analysis
}
