% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.list2tab.R
\name{convert.list2tab}
\alias{convert.list2tab}
\title{Check Ecopath table function}
\usage{
convert.list2tab(diagn.list)
}
\arguments{
\item{diagn.list}{is the list object returned by the create.ETdiagnosis function.}
}
\description{
This function enables the verification of input tables based on EwE data and used in the EcoTroph routine. A template is provided in the example: data(ecopath_guinee).
}
\examples{
data(ecopath_guinee)
\donttest{Liste=create.ETdiagnosis(create.ETmain(ecopath_guinee))
Tab=convert.list2tab(Liste)
}
}
