% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mf.diagnosis.R
\name{mf.diagnosis}
\alias{mf.diagnosis}
\title{Effort multiplier diagnosis}
\usage{
mf.diagnosis(
  x,
  ET_Main,
  catch.list,
  TL_out,
  fleet,
  n.fleet,
  Fish_mort_ref,
  Fish_mort_acc_ref,
  B.Input,
  Beta,
  TopD,
  FormD,
  TLpred,
  n.TL,
  range.TLpred,
  lim.high.TL,
  range.highTL
)
}
\arguments{
\item{x}{is the list object returned by the create.ETdiagnosis function.}

\item{ET_Main}{is the list object returned by the create.ETmain function.}

\item{catch.list}{catches for each fleet}

\item{TL_out}{Maximum TL to consider}

\item{fleet}{List of available fleet}

\item{n.fleet}{Number of fleet}

\item{Fish_mort_ref}{Fishing mortality}

\item{Fish_mort_acc_ref}{Accessible Fisshing mortality}

\item{B.Input}{Biomass}

\item{Beta}{Beta parameters of the formula}

\item{TopD}{Parameters of the formula}

\item{FormD}{Parameters of the formula}

\item{TLpred}{Trophic level for predators}

\item{n.TL}{Number of trophic levels}

\item{range.TLpred}{Range of trophic level for predators}

\item{lim.high.TL}{Limit for high trophic levels}

\item{range.highTL}{Range for High trophic levels}
}
\description{
Effort multiplier diagnosis
}
