% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.smooth.R
\name{plot.smooth}
\alias{plot.smooth}
\title{plot.smooth is used to plot the Smooth function. This function enables the user to see the TL distributions around their mean trophic levels.}
\usage{
\method{plot}{smooth}(x, ...)
}
\arguments{
\item{x}{is the table returned by the create.smooth function.}

\item{\dots}{plot other arguments}
}
\description{
plot.smooth is used to plot the Smooth function. This function enables the user to see the TL distributions around their mean trophic levels.
}
\examples{
data(ecopath_guinee)
plot(create.smooth(ecopath_guinee))
plot(create.smooth(ecopath_guinee,smooth_type=2))
}
\seealso{
create.smooth function to create the Smooth, Transpose to calculate the data transposition into trophic spectra.
}
