% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regPB.ac.R
\name{regPB.ac}
\alias{regPB.ac}
\title{function used to compute pB for the higest trophic levels and accessible biomass
 @param compteur
 @param pb.mf
 @param TL_out
 @param range.highTL}
\usage{
regPB.ac(compteur, pb.mf.ac, TL_out, range.highTL)
}
\description{
function used to compute pB for the higest trophic levels and accessible biomass
 @param compteur
 @param pb.mf
 @param TL_out
 @param range.highTL
}
