% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regPB.R
\name{regPB}
\alias{regPB}
\title{function used to compute pB for the higest trophic levels
 @param compteur
 @param pb.mf
 @param TL_out
 @param range.highTL}
\usage{
regPB(compteur, pb.mf, TL_out, range.highTL)
}
\description{
function used to compute pB for the higest trophic levels
 @param compteur
 @param pb.mf
 @param TL_out
 @param range.highTL
}
