% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.table.R
\name{check.table}
\alias{check.table}
\title{convert.list2tab enables to convert the list object returned by the create.ETdiagnosis function into a list of data.frames. These data.frames contain calculated variables by TL class and combinations of effort multipliers.}
\usage{
check.table(ecopath)
}
\arguments{
\item{ecopath}{is the input table used in ET (possibly based on Ecopath data). The different variables are the group name, its trophic level, biomass, production, catches, omnivory index and accessibility (fraction of the group that can be catch assuming an infinite fishing effort).}
}
\description{
convert.list2tab enables to convert the list object returned by the create.ETdiagnosis function into a list of data.frames. These data.frames contain calculated variables by TL class and combinations of effort multipliers.
}
\examples{
data(ecopath_guinee)
check.table(ecopath_guinee)
}
