% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.ecopath.model.R
\name{read.ecopath.model}
\alias{read.ecopath.model}
\title{Input data import function (from an xml file)}
\usage{
read.ecopath.model(filename)
}
\arguments{
\item{filename}{is the address of the file the user wants to import.}
}
\value{
This function returns a data.frame containing all the column needed to run the EcoTroph R package.
}
\description{
This function loads input data from an xml file created by the user, or exported from the EwE EcoTroph plug-in, or from a web service associated to a database populated with parameters of several EwE models.
}
\seealso{
check.table to control the reliability of the dataset.
}
