% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regPB.ac.R
\name{regPB.ac}
\alias{regPB.ac}
\title{function used to compute pB for the higest trophic levels and accessible biomass}
\usage{
regPB.ac(compteur, pb.mf.ac, TL_out, range.highTL)
}
\arguments{
\item{compteur}{counter}

\item{pb.mf.ac}{Effort multiplier}

\item{TL_out}{Trophic level out of the scope}

\item{range.highTL}{Range of high trophic level}
}
\description{
function used to compute pB for the higest trophic levels and accessible biomass
}
