% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regPB.R
\name{regPB}
\alias{regPB}
\title{function used to compute pB for the higest trophic levels}
\usage{
regPB(compteur, pb.mf, TL_out, range.highTL)
}
\arguments{
\item{compteur}{counter}

\item{pb.mf}{Effort multiplier vector}

\item{TL_out}{Trophic level out of the scope}

\item{range.highTL}{Range of High Trophic level}
}
\description{
function used to compute pB for the higest trophic levels
}
