% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saturation.R
\name{saturation}
\alias{saturation}
\title{Sigma Saturation Function}
\usage{
saturation(sigma_inf = NULL, coeff = NULL, pas = NULL)
}
\arguments{
\item{sigma_inf}{defines the value of the curve's plateau.}

\item{coeff}{defines the value of the slope.}

\item{pas}{defines the splitting of the TL classes.}
}
\value{
saturation returns a vector of values for the sigma used in the create.smooth function.
}
\description{
This function enables an other calculation for the sigma of the create.smooth function. Sigma is calculated on the base of a saturation function reflecting a biological reasoning about the variability of the TL within trophic classes: the variability increases with the TL and reaches a plateau after a certain TL.
}
\details{
By default sigma is constant. This function enables an other user choice reflecting a different reasoning.
}
\examples{
plot(saturation())
lines(saturation(0.2))
text(48,0.18,"sigma_inf=0.2")
lines(saturation(coeff=0.5))
text(48,0.35,"coeff=0.5")
}
\seealso{
create.smooth function to create the Smooth, plot.smooth to plot the smooth function.
}
